# agenda-consultas/swagger-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agenda Consultas API",
    "description": "Versao anterior da API REST para gerenciamento de consultas medicas.",
    "contact": {
      "name": "Suporte",
      "email": "suporte@agendaconsultas.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/consultas/paciente/{pacienteId}": {
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Lista consultas por paciente.",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consulta"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consulta"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consulta"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pacientes/{id}": {
      "get": {
        "tags": [
          "Pacientes"
        ],
        "summary": "Busca um paciente pelo id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Paciente"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paciente"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paciente"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pacientes"
        ],
        "summary": "Atualiza um paciente por id (admin).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PacienteUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PacienteUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PacienteUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Pacientes"
        ],
        "summary": "Atualiza parcialmente um paciente por id (admin).",
        "description": "Envie apenas os campos que deseja alterar. Exemplo:\r\n{\r\n  \"telefone\": \"11988887777\"\r\n}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PacientePatchDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PacientePatchDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PacientePatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Pacientes"
        ],
        "summary": "Remove um paciente por id (admin).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Consulta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "pacienteId": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "dataBrasil": {
            "type": "string",
            "nullable": true
          },
          "especialidade": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Paciente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "cpf": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PacientePatchDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "nullable": true
          },
          "cpf": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PacienteUpdateDto": {
        "required": [
          "cpf",
          "email",
          "nome",
          "telefone"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "minLength": 3,
            "type": "string"
          },
          "cpf": {
            "maxLength": 14,
            "minLength": 11,
            "type": "string"
          },
          "telefone": {
            "minLength": 8,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Informe o token JWT no formato: Bearer {seu_token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}